bot-pilot-chat/backend/bot-connectors/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

google-api-python-client==2.149.0
google-auth==2.40.3
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

requests==2.32.3
httpx==0.27.2
python-dateutil==2.8.2

sqlalchemy==2.0.43
pymysql==1.1.1

# Dev Tools
pytest==8.3.2
flake8==7.1.1
black==24.8.0